Initial release of wmflib
This new package is basically a replica of the Spicerack layout.
It contains, for the moment, only the dns module, adapted from Spicerack,
and its tests. The goal of this package is to move functionalities that
do not require high user privileges out of Spicerack, to be re-used
in projects/scripts that need it.

Bug: T257905
Change-Id: I919867a32e9b2a765ee0dc7cdda94b75b37b9e96