dns:  update DNS to support multiple namservers
This allows cookbooks to configure the Dns with multiple nameservers
  dns = Dns(nameserver_address=['91.198.174.239', '208.80.153.231']),
and thus allow users to get authoritative answers whiles also making use
of DNS failover to account for any on going work on a specific
nameserver while the cookbook is running
  e.g. https://gerrit.wikimedia.org/r/656139

I have also created the following variable to hold the auth server ips.
We could calculate theses at run time by doing walking the nsset from the
root domain.  however NS set changes happen very infrequently and will
require a lot of messaging and lead time as such i think we can forgo
this complexity and KISS by just specifying the static list and updating
it if this ever changes in the future.

Change-Id: I1ffdee91a64ba6ad958fbf8812f53fb001ca598f